Move form mock data into own folder
