+++ source/assets/mock/forms.js
@@ -1,0 +1,6 @@
+import EKB_FORM from './form-ekb';
+import BVL_FORM from './form-bvl';
+
+const forms = [BVL_FORM, EKB_FORM];
+
+export default forms;
