+++ source/components/organisms/FormAgent/FormAgent.js
@@ -6,7 +6,7 @@
 import React, { Component } from 'react';
 import validator from 'validator';
 import EventHandler, { EVENT_USER_MESSAGE } from '../../../helpers/EventHandler';
-import forms from '../../../assets/forms.js';
+import forms from '../../../assets/mock/forms.js';
 import ChatBubble from '../../atoms/ChatBubble';
 import ChatDivider from '../../atoms/ChatDivider/ChatDivider';
 import StorageService, { USER_KEY } from '../../../services/StorageService';
